Wrote more tests on command line parsing (main functionality), and refactored out of Program.Main
